pyproject.toml
[project]
name = "coros-mcp"
version = "0.1.0"
description = "MCP server for Coros sleep and HRV data"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.9",
    "httpx>=0.27",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
coros-mcp = "server:main"
